+++ bocadillo/__init__.py
@@ -1,8 +1,12 @@
 from .applications import App
+from .config import settings, SettingsError, configure
+from .compat import ExpectedAsync
 from .errors import HTTPError
 from .middleware import ASGIMiddleware, Middleware
 from .injection import discover_providers, provider, useprovider
+from .plugins import plugin, get_plugins
 from .recipes import Recipe
+from .redirection import Redirect
 from .request import ClientDisconnect, Request
 from .response import Response
 from .sse import server_event
@@ -11,4 +15,4 @@ from .templates import Templates
 from .views import view
 from .websockets import WebSocket, WebSocketDisconnect
 
-__version__ = "0.13.2"
+__version__ = "0.13.3"