+++ bocadillo/errors.py
@@ -1,8 +1,8 @@
 from http import HTTPStatus
-from typing import Any, Dict, Optional, Type, Union
+import typing
 
 from .app_types import _E, ErrorHandler, HTTPApp
-from .compat import call_async
+from .compat import check_async
 from .request import Request
 from .response import Response
 
@@ -26,7 +26,9 @@ class HTTPError(Exception):
 
     __slots__ = ("_status", "detail")
 
-    def __init__(self, status: Union[int, HTTPStatus], detail: Any = ""):
+    def __init__(
+        self, status: typing.Union[int, HTTPStatus], detail: typing.Any = ""
+    ):
         if isinstance(status, int):
             status = HTTPStatus(  # pylint: disable=no-value-for-parameter
                 status
@@ -68,7 +70,7 @@ class ServerErrorMiddleware(HTTPApp):
     def __init__(self, app: HTTPApp, handler: ErrorHandler) -> None:
         self.app = app
         self.handler = handler
-        self.exception: Optional[BaseException] = None
+        self.exception: typing.Optional[BaseException] = None
 
     def raise_if_exception(self):
         if self.exception is not None:
@@ -79,9 +81,7 @@ class ServerErrorMiddleware(HTTPApp):
             res = await self.app(req, res)
         except BaseException as exc:
             self.exception = exc
-            await call_async(  # type: ignore
-                self.handler, req, res, HTTPError(500)
-            )
+            await self.handler(req, res, HTTPError(500))
             return res
         else:
             return res
@@ -97,15 +97,23 @@ class HTTPErrorMiddleware(HTTPApp):
 
     def __init__(self, app: HTTPApp) -> None:
         self.app = app
-        self._exception_handlers: Dict[Type[BaseException], ErrorHandler] = {}
+        self._exception_handlers: typing.Dict[
+            typing.Type[BaseException], ErrorHandler
+        ] = {}
 
     def add_exception_handler(
-        self, exception_class: Type[_E], handler: ErrorHandler
+        self, exception_class: typing.Type[_E], handler: ErrorHandler
     ) -> None:
-        assert issubclass(exception_class, BaseException)
+        assert issubclass(
+            exception_class, BaseException
+        ), f"expected an exception class, not {type(exception_class)}"
+        check_async(
+            handler,
+            reason=f"error handler '{handler.__name__}' must be asynchronous",
+        )
         self._exception_handlers[exception_class] = handler
 
-    def _get_exception_handler(self, exc: _E) -> Optional[ErrorHandler]:
+    def _get_exception_handler(self, exc: _E) -> typing.Optional[ErrorHandler]:
         for cls, handler in self._exception_handlers.items():
             if issubclass(type(exc), cls):
                 return handler
@@ -129,7 +137,7 @@ class HTTPErrorMiddleware(HTTPApp):
                 handler = self._get_exception_handler(exc)
                 if handler is None:
                     raise exc from None
-                response = call_async(handler, req, res, exc)  # type: ignore
+                response = handler(req, res, exc)
 
             if not has_error:
                 assert res is not None