+++ bocadillo/routing.py
@@ -9,40 +9,29 @@
 """
 
 import inspect
-from typing import (
-    Any,
-    Callable,
-    Dict,
-    Generic,
-    NoReturn,
-    Optional,
-    Tuple,
-    Type,
-    TypeVar,
-)
+import typing
 
 from starlette.websockets import WebSocketClose
 
 from . import views
 from .app_types import HTTPApp, Receive, Scope, Send
 from .errors import HTTPError
-from .injection import consumer
-from .redirection import Redirection
+from .redirection import Redirect
 from .request import Request
 from .response import Response
 from .urlparse import Parser
-from .views import AsyncHandler, HandlerDoesNotExist, View
+from .views import Handler, HandlerDoesNotExist, View
 from .websockets import WebSocket, WebSocketView
 
 # Route generic types.
-_R = TypeVar("_R", bound="BaseRoute")  # route
-_V = TypeVar("_V")  # view
+_R = typing.TypeVar("_R", bound="BaseRoute")  # route
+_V = typing.TypeVar("_V")  # view
 
 
 # Utilities.
 
 
-class RouteMatch(Generic[_R]):  # pylint: disable=unsubscriptable-object
+class RouteMatch(typing.Generic[_R]):  # pylint: disable=unsubscriptable-object
     """Represents a match between an URL path and a route.
 
     # Parameters
@@ -60,7 +49,7 @@ class RouteMatch(Generic[_R]):  # pylint: disable=unsubscriptable-object
 # Base classes.
 
 
-class BaseRoute(Generic[_V]):
+class BaseRoute(typing.Generic[_V]):
     """The base route class.
 
     This is referenced as `_R` in the rest of this module.
@@ -82,20 +71,7 @@ class BaseRoute(Generic[_V]):
     def pattern(self) -> str:
         return self._parser.pattern
 
-    def url(self, **kwargs) -> str:
-        """Return the full URL path for the given route parameters.
-
-        # Parameters
-        kwargs (dict): route parameters.
-
-        # Returns
-        url (str):
-            A full URL path obtained by formatting the route pattern with
-            the provided route parameters.
-        """
-        return self.pattern.format(**kwargs)
-
-    def parse(self, path: str) -> Optional[dict]:
+    def parse(self, path: str) -> typing.Optional[dict]:
         """Parse an URL path against the route's URL pattern.
 
         # Returns
@@ -106,7 +82,7 @@ class BaseRoute(Generic[_V]):
         return self._parser.parse(path)
 
     @classmethod
-    def normalize(cls, view: Any) -> _V:
+    def normalize(cls, view: typing.Any) -> _V:
         """Perform any conversion necessary to return a proper view object.
 
         Not implemented.
@@ -119,13 +95,13 @@ class BaseRoute(Generic[_V]):
         return cls(view=view, pattern=pattern, **kwargs)
 
     @classmethod
-    def create(cls, view: Any, pattern: str, **kwargs) -> _R:
+    def create(cls, view: typing.Any, pattern: str, **kwargs) -> _R:
         """Normalize a view and build and a route instance."""
         view: _V = cls.normalize(view)
         return cls.build(view, pattern=pattern, **kwargs)
 
 
-class BaseRouter(Generic[_R, _V]):
+class BaseRouter(typing.Generic[_R, _V]):
     """The base router class.
 
     # Attributes
@@ -135,10 +111,10 @@ class BaseRouter(Generic[_R, _V]):
 
     __slots__ = ("routes", "route_class")
 
-    route_class: Type[_R]
+    route_class: typing.Type[_R]
 
     def __init__(self):
-        self.routes: Dict[str, _R] = {}
+        self.routes: typing.Dict[str, _R] = {}
 
     def _get_key(self, route: _R) -> str:
         # Return the key at which `route` should be stored internally.
@@ -148,7 +124,7 @@ class BaseRouter(Generic[_R, _V]):
         """Register a route."""
         self.routes[self._get_key(route)] = route
 
-    def route(self, *args, **kwargs) -> Callable[[Any], _R]:
+    def route(self, *args, **kwargs) -> typing.Callable[[typing.Any], _R]:
         """Register a route by decorating a view.
 
         The decorated function or class will be converted to a proper view using
@@ -161,13 +137,13 @@ class BaseRouter(Generic[_R, _V]):
             along with the normalized view.
         """
 
-        def decorate(view: Any) -> _R:
+        def decorate(view: typing.Any) -> _R:
             route = self.route_class.create(view, *args, **kwargs)
             self.add_route(route)
 
         return decorate
 
-    def match(self, path: str) -> Optional[RouteMatch[_R]]:
+    def match(self, path: str) -> typing.Optional[RouteMatch[_R]]:
         """Attempt to match an URL path against one of the registered routes.
 
         # Parameters
@@ -209,7 +185,7 @@ class HTTPRoute(BaseRoute[View]):
         self.name = name
 
     @classmethod
-    def normalize(cls, view: Any) -> View:
+    def normalize(cls, view: typing.Any) -> View:
         """Build a #::bocadillo.views#View object.
 
         The input, free-form `view` object is converted using the following
@@ -278,7 +254,7 @@ class HTTPRoute(BaseRoute[View]):
         method: str = req.method.lower()
 
         try:
-            handler: AsyncHandler = self.view.get_handler(method)
+            handler: Handler = self.view.get_handler(method)
         except HandlerDoesNotExist as e:
             raise HTTPError(405) from e
 
@@ -308,8 +284,8 @@ class HTTPRouter(HTTPApp, BaseRouter[HTTPRoute, View]):
 
         try:
             await match.route(req, res, **match.params)
-        except Redirection as redirection:
-            res = redirection.response
+        except Redirect as exc:
+            res = exc.response
 
         return res
 
@@ -338,7 +314,7 @@ class WebSocketRoute(BaseRoute[WebSocketView]):
         self._ws_kwargs = kwargs
 
     @classmethod
-    def normalize(cls, view: Any) -> WebSocketView:
+    def normalize(cls, view: typing.Any) -> WebSocketView:
         return WebSocketView(view)
 
     async def __call__(
@@ -409,7 +385,7 @@ class RoutingMixin:
         value_type: str = None,
         receive_type: str = None,
         send_type: str = None,
-        caught_close_codes: Tuple[int] = None,
+        caught_close_codes: typing.Tuple[int] = None,
     ):
         """Register a WebSocket route by decorating a view.
 
@@ -430,56 +406,3 @@ class RoutingMixin:
             send_type=send_type,
             caught_close_codes=caught_close_codes,
         )
-
-    def url_for(self, name: str, **kwargs) -> str:
-        """Build the full URL path for a named #::bocadillo.routing#HTTPRoute.
-
-        # Parameters
-        name (str): the name of the route.
-        kwargs (dict): route parameters.
-
-        # Returns
-        url (str): an URL path.
-
-        # Raises
-        HTTPError(404) : if no route exists for the given `name`.
-        """
-        route = self.http_router.routes.get(name)
-        if route is None:
-            raise HTTPError(404)
-        return route.url(**kwargs)
-
-    def redirect(
-        self,
-        *,
-        name: str = None,
-        url: str = None,
-        permanent: bool = False,
-        **kwargs,
-    ) -> NoReturn:
-        """Redirect to another #::bocadillo.routing#HTTPRoute.
-
-        This is only meant to be used inside an HTTP view.
-
-        # Parameters
-        name (str): name of the route to redirect to.
-        url (str):
-            URL of the route to redirect to (required if `name` is omitted).
-        permanent (bool):
-            If `False` (the default), returns a temporary redirection (302).
-            If `True`, returns a permanent redirection (301).
-        kwargs (dict):
-            Route parameters.
-
-        # Raises
-        Redirection:
-            an exception that will be caught to trigger a redirection.
-
-        # See Also
-        - [Redirecting](../guides/http/redirecting.md)
-        """
-        if name is not None:
-            url = self.url_for(name=name, **kwargs)
-        else:
-            assert url is not None, "url is expected if no route name is given"
-        raise Redirection(url=url, permanent=permanent)